test: move TcpProxy to tests/tcp_proxy.py to fix collection
tests/unit/test_tcp_proxy.py, added in d99dc460, imported its subject
(TcpProxy) from tests/integration/standard/test_client_routes.py, which
transitively imports tests/integration/__init__.py. That module guards its
ccmlib imports with try/except ImportError, but then unconditionally
declares `class Cassandra41CCMCluster(CCMCluster)` at module level, so on
any environment without ccmlib installed the import fails with:

    NameError: name 'CCMCluster' is not defined

This broke test collection consistently on the windows-2022 job, where
ccmlib is absent. The latent defect in tests/integration/__init__.py
predates d99dc460; that commit merely became the first unit test to import
tests.integration and thus the first to expose it.

TcpProxy is a plain socket-based helper -- it depends only on socket,
select and threading, and needs neither CCM nor a running
Cassandra/Scylla cluster -- so it does not belong behind that import.
Move it verbatim into a new tests/tcp_proxy.py and import it from both
call sites:

- tests/integration/standard/test_client_routes.py now imports TcpProxy
  from tests.tcp_proxy; its `select` and `socket` imports, used only by
  the moved class, are dropped.
- tests/unit/test_tcp_proxy.py imports from tests.tcp_proxy and no longer
  needs its os.environ.setdefault("CASSANDRA_VERSION", ...) shim, which
  existed solely to get tests.integration's module-level version parsing
  to succeed. The shim and the docstring paragraph explaining it are
  removed.

The class body is byte-identical to the original; only the new module's
license header, docstring and imports are new. No driver code is touched
and no test behavior changes.

Validation:
- pytest tests/unit/test_tcp_proxy.py: 2 passed with neither
  CASSANDRA_VERSION nor SCYLLA_VERSION set, i.e. the unit test no longer
  imports tests.integration at all.
- tests/integration/standard/test_client_routes.py compiles clean with no
  imports left unused.

Fixes: scylladb/python-driver#965
